Actually rename namespace
